// frontend/lib/generated/fastapi/schemas/memberOverviewSummaryPayload.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberOverviewSummaryPoint } from './memberOverviewSummaryPoint';

export interface MemberOverviewSummaryPayload {
  rangeDays: number;
  totalViews: number;
  uniqueVisitors: number;
  upvotesInRange: number;
  viewsOverTime: MemberOverviewSummaryPoint[];
}